% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial.R
\name{financial}
\alias{financial}
\title{Randomly Generated Data}
\format{
A data frame with 100 observations of 6 variables called financial
}
\usage{
financial()
}
\value{
A data frame containing the loaded "financial" data from the .rds file.
\itemize{
\item \code{Dcollege}: Dummy variable indicating college education.
\item \code{Dunemp}: Dummy variable indicating unemployment.
\item \code{Totalincome}: Total income of each observation.
\item \code{poor_liq}: Dummy variable indicating liquid poverty.
\item \code{w}: Weights for each observation.
\item \code{n}: Identifier for observations.
}
}
\description{
This dataset contains 100 observations of 6 variables. The data was
generated randomly for the purpose of exemplifying a database that could
potentially be used with this function.
}
\examples{
data(financial)
head(financial)


}
\keyword{dataset}
\keyword{example}
