% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\docType{data}
\name{social}
\alias{social}
\title{Randomly Generated Data}
\format{
A data frame with 200 observations of 6 variables called social
}
\usage{
social()
}
\value{
A data frame containing the loaded "social" data from the .rds file.
\itemize{
\item \code{Dcollege}: Dummy variable indicating college education.
\item \code{Dunemp}: Dummy variable indicating unemployment.
\item \code{Totalincome}: Total income of each observation.
\item \code{reg}:Variable indicating the region of the observation.
\item \code{w}: Weights for each observation.
\item \code{n}: Identifier for observations.
}
#' @examples
data(social)
head(social)
}
\description{
This dataset contains 200 observations of 6 variables. The data was
generated randomly for the purpose of exemplifying a database that could
potentially be used with this function.
}
\keyword{dataset}
\keyword{example}
