% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_gce.R
\name{summary.kl}
\alias{summary.kl}
\title{Summary}
\usage{
\method{summary}{kl}(object, ...)
}
\arguments{
\item{object}{The output obtained from ei_gce}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
This summary function returns the Kullback-Leibler divergence value and the last iteration in the optimization process.
A dataframe with the means of the estimations for each characteristic j with the predictions the probabilities and the error estimated.
A dataframe with the lambda estimated for each k.
\itemize{
\item \code{Iterations}:Indicates the times the objective function and the gradient has been evaluated during the optimization process
\item \code{divergencekl value}:The Kullback-Leibler divergence value resulting from the optimization.
\item \code{mean_estimations}:The weighted mean of predictions, probabilities, and the errors for each category j of the variable y
\item \code{lambda}:The estimated lambda values.
}
}
\description{
This function provides a summary of the output obtained with the function ei_gce.
}
