% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_gme.R
\name{summary.shannon}
\alias{summary.shannon}
\title{Summary}
\usage{
\method{summary}{shannon}(object, ...)
}
\arguments{
\item{object}{The output obtained from ei_gme}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
This summary function returns the entropy value and the last iteration in the optimization process.
A dataframe with the means of the estimations for each characteristic j with the predictions the probabilities and the error estimated.
A dataframe with the lambda estimated for each k.
\itemize{
\item \code{Iterations}:Indicates the times the objective function and the gradient has been evaluated during the optimization process
\item \code{Entropy value}:The value of entropy resulting from the optimization.
\item \code{mean_estimations}: The predictions, p_dual, and the error for each category j of the variable y
\item \code{lambda}:The estimated lambda values.
}
}
\description{
This function provides a summary of the output obtained with the function ei_gme.
}
